requirements.txt
requests
python-dotenv
psycopg2-binary